requirements.txt
Django>=3.0
django-bootstrap4>=3.0.0
gunicorn==20.1.0
mysqlclient==2.2.4
requests==2.30.0
dj-database-url==2.0.0
psycopg2-binary==2.9.6
